Fix double ups of results for the same test
This was an issue with the re-syncing of the results where the line ID
was given a +1 for some reason. This meant that re-synched results had
an ID that did not match their line number, and from there they were no
longer kept in sync properly. Removed the +1 thing and am hoping this
fixes it for good.