Handle parsing results for files that are not open
Running all specs in a code base may mean that results are interpreted
for files that have not been opened in the editor.

When the file is not open we are not able to get the line content of the
example, so when the presenter attempts to sync the line it ends up
getting deleted. We could try and open the file somehow to read it (
might be the better option), or we could just not grab the line content.
Which is what we are doing in this commit as it the simplest option in
terms of making what is already there just work. Unsure if there will be
any weird side effects of this.