Minitest supports running from context
Supported by codelens only. Use the context name with the -n option to
get minitest to run all tests for that context.

Minitest doesn't tell us which line numbers passed, so we need to figure
it out for ourselves and if the test results don't mention them then
they didn't fail. To do this we get the line for the context and
remember the indentation. Then we go down the file finding test lines
with our minitest parser regex. If we find a line with less or the same
indentation as the line the context had, we stop searching and assume
that we've found the end of the context.