Fix issue with error highlights in rspec
Issue was that rspec error backtraces was using relative paths which
was confusing the extension. Now relative paths are supported so the
error highlights show up again.

Also add some coloured underlines to the test names for fun.

And update all the dependencies.

And disable some tests sadly. Tried to debug them but I don't know how.
Vscode docs have a launch config for this but it doesn't work. Didn't
work when I created a new blank extension either. So I'm guessing that
the docs are just out of date and I don't have time to figure out how to
attach a debugger without help from docs.